delete validation

+++ controller/projectController.js
@@ -57,7 +57,14 @@ const deleteProject = async (req, res) => {
     if (!project) {
       return res.status(404).json({ error: "Project not found!" });
     }
+    
+    if(project.user.toString() !== req.user._id.toString()){
+      return res
+        .status(401)
+        .json({ error: "Unauthorised to deleted the project" });
+    }
 
+    
     await Project.findByIdAndDelete(req.params.id);
     await User.findByIdAndUpdate(project.user, {
       $pull: { projects: project._id },
